# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/phones": {
      "get": {
        "operationId": "PhonesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start date (shop-local)",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive end date (shop-local)",
            "schema": {
              "example": "2026-08-31",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "IN_STOCK",
                "SOLD"
              ]
            }
          },
          {
            "name": "condition",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "NEW",
                "USED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Match on name or IMEI",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "name",
                "purchasePrice"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "DESC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List phones (filter/search/paginate)",
        "tags": [
          "phones"
        ]
      },
      "post": {
        "operationId": "PhonesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePhoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a phone (IN_STOCK)",
        "tags": [
          "phones"
        ]
      }
    },
    "/phones/{id}": {
      "get": {
        "operationId": "PhonesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one phone",
        "tags": [
          "phones"
        ]
      },
      "patch": {
        "operationId": "PhonesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePhoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a phone",
        "tags": [
          "phones"
        ]
      },
      "delete": {
        "operationId": "PhonesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft-delete a phone (only while IN_STOCK)",
        "tags": [
          "phones"
        ]
      }
    },
    "/phones/{id}/label": {
      "get": {
        "operationId": "PhonesController_label",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneLabelDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Spec-card label payload (never contains a price)",
        "tags": [
          "phones"
        ]
      }
    },
...
    "/sales": {
      "get": {
        "operationId": "SalesController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Inclusive start date (shop-local)",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Inclusive end date (shop-local)",
            "schema": {
              "example": "2026-08-31",
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PHONE",
                "ACCESSORY"
              ]
            }
          },
          {
            "name": "isDebt",
            "required": false,
            "in": "query",
            "description": "Only sales that carry a debt",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Match sale code or product name/imei",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List sales (type/isDebt/date/search)",
        "tags": [
          "sales"
        ]
      }
    },
    "/sales/{id}": {
      "get": {
        "operationId": "SalesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get one sale with items, product snapshot & debt",
        "tags": [
          "sales"
        ]
      }
    },
    "/sales/phone": {
      "post": {
        "operationId": "SalesController_sellPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellPhoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Sell a phone (optional debt)",
        "tags": [
          "sales"
        ]
      }
    },
    "/sales/accessory": {
      "post": {
        "operationId": "SalesController_sellAccessory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellAccessoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Sell accessories (optional debt)",
        "tags": [
          "sales"
        ]
      }
    },
    "/sales/{id}/return": {
      "post": {
        "operationId": "SalesController_createReturn",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReturnDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Return sold units from a sale",
        "tags": [
          "sales"
        ]
      }
    },
    "/sales/{id}/returns": {
      "get": {
        "operationId": "SalesController_listReturns",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SaleReturnResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List returns for a sale",
        "tags": [
          "sales"
        ]
      }
    },
...
  },
  "info": {
    "title": "Phone & Accessory Shop API",
    "description": "Backend for a small retail shop: phones, accessories, sales, debts, expenses and statistics. Multi-tenant — all data scoped by shop.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
      "CreatePhoneDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Samsung Galaxy A54 5G"
          },
          "imei": {
            "type": "string",
            "example": "356938035643809",
            "description": "digits only, 10–20 chars"
          },
          "purchasePrice": {
            "type": "number",
            "example": 2500000,
            "description": "What the shop paid (UZS)"
          },
          "listPrice": {
            "type": "number",
            "example": 3000000,
            "description": "Internal reference only"
          },
          "condition": {
            "type": "string",
            "enum": [
              "NEW",
              "USED"
            ]
          },
          "ramGb": {
            "type": "number",
            "example": 8
          },
          "storageGb": {
            "type": "number",
            "example": 256
          },
          "imageUrl": {
            "type": "string",
            "example": "/uploads/abc.webp"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "imei",
          "purchasePrice"
        ]
      },
      "PhoneResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "imei": {
            "type": "string"
          },
          "purchasePrice": {
            "type": "number"
          },
          "listPrice": {
            "type": "object",
            "nullable": true
          },
          "condition": {
            "type": "string",
            "enum": [
              "NEW",
              "USED"
            ],
            "nullable": true
          },
          "ramGb": {
            "type": "object",
            "nullable": true
          },
          "storageGb": {
            "type": "object",
            "nullable": true
          },
          "imageUrl": {
            "type": "object",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "IN_STOCK",
              "SOLD"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "imei",
          "purchasePrice",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdatePhoneDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Samsung Galaxy A54 5G"
          },
          "imei": {
            "type": "string",
            "example": "356938035643809",
            "description": "digits only, 10–20 chars"
          },
          "purchasePrice": {
            "type": "number",
            "example": 2500000,
            "description": "What the shop paid (UZS)"
          },
          "listPrice": {
            "type": "number",
            "example": 3000000,
            "description": "Internal reference only"
          },
          "condition": {
            "type": "string",
            "enum": [
              "NEW",
              "USED"
            ]
          },
          "ramGb": {
            "type": "number",
            "example": 8
          },
          "storageGb": {
            "type": "number",
            "example": 256
          },
          "imageUrl": {
            "type": "string",
            "example": "/uploads/abc.webp"
          },
          "note": {
            "type": "string"
          }
        }
      },
      "PhoneLabelDto": {
        "type": "object",
        "properties": {
          "shopName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "memory": {
            "type": "object",
            "nullable": true,
            "example": "8 GB / 256 GB",
            "description": "Omitted when both ram and storage are null"
          },
          "condition": {
            "type": "object",
            "nullable": true
          },
          "imei": {
            "type": "string"
          },
          "labelFooter": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "shopName",
          "name",
          "imei"
        ]
      },
...
      "SaleProductSnapshotDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "imei": {
            "type": "object",
            "nullable": true
          },
          "memory": {
            "type": "object",
            "nullable": true,
            "example": "8 GB / 256 GB"
          },
          "imageUrl": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "SaleItemResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "itemType": {
            "type": "string",
            "enum": [
              "PHONE",
              "ACCESSORY"
            ]
          },
          "quantity": {
            "type": "number"
          },
          "unitPrice": {
            "type": "number"
          },
          "costPrice": {
            "type": "number"
          },
          "lineTotal": {
            "type": "number"
          },
          "returnedQuantity": {
            "type": "number"
          },
          "product": {
            "$ref": "#/components/schemas/SaleProductSnapshotDto"
          }
        },
        "required": [
          "id",
          "itemType",
          "quantity",
          "unitPrice",
          "costPrice",
          "lineTotal",
          "returnedQuantity",
          "product"
        ]
      },
      "SaleDebtSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "customerName": {
            "type": "string"
          },
          "customerPhone": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "dueDate": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "isOverdue": {
            "type": "boolean"
          },
          "daysOverdue": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "customerName",
          "customerPhone",
          "amount",
          "dueDate",
          "status",
          "isOverdue",
          "daysOverdue"
        ]
      },
      "SaleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PHONE",
              "ACCESSORY"
            ]
          },
          "totalAmount": {
            "type": "number"
          },
          "paidAmount": {
            "type": "number"
          },
          "debtAmount": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETED",
              "PARTIALLY_RETURNED",
              "RETURNED"
            ]
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "soldAt": {
            "format": "date-time",
            "type": "string"
          },
          "profit": {
            "type": "number",
            "description": "Profit net of returns: sum (unitPrice - costPrice) * (qty - returnedQty)"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleItemResponseDto"
            }
          },
          "debt": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/SaleDebtSummaryDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "code",
          "type",
          "totalAmount",
          "paidAmount",
          "debtAmount",
          "status",
          "soldAt",
          "profit",
          "items"
        ]
      },
      "DebtInputDto": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "example": 500000,
            "description": "Amount still owed"
          },
          "dueDate": {
            "type": "string",
            "example": "2026-08-15",
            "description": "Due date (shop-local)"
          },
          "customerName": {
            "type": "string",
            "example": "Ali Valiyev"
          },
          "customerPhone": {
            "type": "string",
            "example": "998901234567",
            "description": "Normalised to 998XXXXXXXXX"
          }
        },
        "required": [
          "amount",
          "dueDate",
          "customerName",
          "customerPhone"
        ]
      },
      "SellPhoneDto": {
        "type": "object",
        "properties": {
          "phoneId": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "example": 3000000,
            "description": "Negotiated sale price"
          },
          "note": {
            "type": "string"
          },
          "debt": {
            "$ref": "#/components/schemas/DebtInputDto"
          }
        },
        "required": [
          "phoneId",
          "price"
        ]
      },
      "SellAccessoryDto": {
        "type": "object",
        "properties": {
          "accessoryId": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "example": 2
          },
          "unitPrice": {
            "type": "number",
            "example": 25000,
            "description": "Defaults to the accessory salePrice when omitted"
          },
          "note": {
            "type": "string"
          },
          "debt": {
            "$ref": "#/components/schemas/DebtInputDto"
          }
        },
        "required": [
          "accessoryId",
          "quantity"
        ]
      },
      "CreateReturnDto": {
        "type": "object",
        "properties": {
          "saleItemId": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "example": 1,
            "description": "Units to return"
          },
          "amount": {
            "type": "number",
            "description": "Refund amount. Defaults to the proportional sold amount; may be lower, never higher."
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "saleItemId",
          "quantity",
          "reason"
        ]
      },
      "SaleReturnResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "saleItemId": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "amount": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "saleItemId",
          "quantity",
          "amount",
          "reason",
          "createdAt"
        ]
      },
...
    }
  }
}